+++ extensions/youtube-downloader/CHANGELOG.md
@@ -1,5 +1,9 @@
 # YouTube Downloader Changelog
 
+## [Fixes] - 2025-02-03
+
+- Fixed error: Unable to get selected text from frontmost application
+
 ## [Insert active tab URL] - 2025-02-02
 
 - If the raycast browser extension is installed, the extension will automatically insert the active tab URL into the input field